# Reduce GPU memory usage in CacheDataloader