--- a/nerfstudio/data/datasets/sdf_dataset.py
+++ b/nerfstudio/data/datasets/sdf_dataset.py
@@ -34,6 +34,8 @@ class SDFDataset(InputDataset):
         scale_factor: The scaling factor for the dataparser outputs.
     """
 
+    exclude_batch_keys_from_device = InputDataset.exclude_batch_keys_from_device + ["depth", "normal"]
+
     def __init__(self, dataparser_outputs: DataparserOutputs, scale_factor: float = 1.0):
         super().__init__(dataparser_outputs, scale_factor)