@tancik 2023-04-13
If you have a lidar iphone, you can create a capture with polycam that has depth.