@jkulhanek 2023-04-13
Unfortunately, I don’t have a lidar iphone. I guess I can run a monocular depth prediction model