@tancik 2023-05-18
I'd be curious to hear @machenmusik perspective on this PR since he has looked more into dataloader/device stuff.