@machenmusik 2023-05-20
> I'd be curious to hear @machenmusik perspective on this PR since he has looked more into dataloader/device stuff.

Stymied by Windows and torch 2 issues at the moment; once that is cleared up, I can hopefully try some other variations with this.