@jkulhanek 2023-06-05
@machenmusik  thanks! Do you, by any chance, also have GPU memory usage stats?