@machenmusik 2023-06-05
> @machenmusik thanks! Do you, by any chance, also have GPU memory usage stats?

No sorry, huge model saturates this machine anyway so wouldn't see anything there really